Create modules 'renderSideBar' and 'renderTodos'
